// src/hooks/homeFetchTypes.d.ts
type TListMovie = {
	adult: boolean;
	backdrop_path: string;
	genre_ids: number[];
	id: number;
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	release_date: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}

export type TListMovieDetails = {
  page: number;
  results: TListMovie[];
  total_pages: number;
  total_results: number;
}